moved extension api functions to one location
