+++ src/popup/components/PRDisplay/RepoSection/NoPullRequest.jsx
@@ -1,7 +1,7 @@
-/* global chrome */
 import React from "react";
 import Box from "@mui/material/Box";
 import Typography from "@mui/material/Typography";
+import { createTab } from "../../../../data/extension";
 
 export default function NoPullRequest({ repo }) {
   return (
@@ -16,9 +16,7 @@ export default function NoPullRequest({ repo }) {
         },
       }}
       onClick={() => {
-        chrome.tabs.create({
-          url: `${repo.url}/pulls`,
-        });
+        createTab(`${repo.url}/pulls`);
       }}
     >
       <Typography variant="caption">No pull requests open</Typography>
